# docs/sa/API_Specification.yaml
openapi: 3.0.3
info:
  title: ServiceDesk CRM API
  description: |
    REST API для системы автоматизации сервисной компании по обслуживанию банкоматов и платёжного оборудования.

    ## Аутентификация
    Все защищённые эндпоинты требуют заголовок `Authorization: Bearer <token>`.
    Токен получается через `POST /auth/login`. Срок действия — 480 минут (8 часов).

    ## Формат ошибок
    ```json
    { "error": "ERROR_CODE", "message": "Описание ошибки", "details": {} }
    ```

    ## Пагинация
    Поддерживается параметрами `?page=1&size=20`.
    Ответ всегда включает `total`, `page`, `size`, `pages`.
  version: "2.0.0"
  contact:
    name: Solution Architect
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: http://localhost/api/v1
    description: Docker Compose (via Nginx)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: TICKET_NOT_FOUND
        message:
          type: string
          example: "Заявка #42 не найдена"
        details:
          type: object
      required: [error, message, details]
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
          description: Текущая страница (начиная с 1)
        size:
          type: integer
          description: Размер страницы
        pages:
          type: integer
          description: Общее количество страниц
      required: [total, page, size, pages]

    # ─── Auth ─────────────────────────────────────────────────────────────────
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
      required: [access_token, token_type]

    # ─── User ─────────────────────────────────────────────────────────────────
    UserRole:
      type: string
      enum: [admin, sales_mgr, svc_mgr, engineer, manager, warehouse, accountant, director]
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        full_name:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
          description: Массив ролей пользователя (union прав)
        phone:
          type: string
          nullable: true
        telegram_chat_id:
          type: string
          nullable: true
        is_active:
          type: boolean
        is_deleted:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, full_name, roles, is_active, is_deleted, created_at, updated_at]
...
    ClientResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        inn:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        assigned_manager_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, name, created_at]
    ClientCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 300
        inn:
          type: string
          maxLength: 12
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        assigned_manager_id:
          type: integer
          nullable: true
      required: [name]
...
    WarrantyStatus:
      type: string
      enum: [valid, expiring_soon, expired]
    EquipmentResponse:
      type: object
      properties:
        id:
          type: integer
        serial_number:
          type: string
        client_id:
          type: integer
        model_id:
          type: integer
        model_name:
          type: string
        location:
          type: string
          nullable: true
        install_date:
          type: string
          format: date
          nullable: true
        warranty_expiry:
          type: string
          format: date
          nullable: true
        warranty_status:
          $ref: '#/components/schemas/WarrantyStatus'
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, serial_number, client_id, model_id, model_name, warranty_status, created_at]
    EquipmentCreate:
      type: object
      properties:
        serial_number:
          type: string
          maxLength: 100
        client_id:
          type: integer
        model_id:
          type: integer
        location:
          type: string
          nullable: true
        install_date:
          type: string
          format: date
          nullable: true
        warranty_expiry:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required: [serial_number, client_id, model_id]
...
  responses:
    Unauthorized:
      description: Токен отсутствует или истёк
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INVALID_TOKEN
            message: "Токен недействителен"
            details: {}
...
    NotFound:
      description: Запись не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт уникальности
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@company.ru
              password: secret123
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: INVALID_CREDENTIALS
                message: "Неверный email или пароль"
                details: {}
  /auth/me:
    get:
      tags: [Auth]
      summary: Текущий пользователь
      responses:
        '200':
          description: Данные текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ═══════════════════════════════════════════════════════════════
  # USERS (Module 8)
  # ═══════════════════════════════════════════════════════════════
...
  /clients:
    get:
      tags: [Клиенты]
      summary: Список клиентов
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Поиск по названию или ИНН
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClientResponse'

    post:
      tags: [Клиенты]
      summary: Создать клиента
      description: admin, sales_mgr, svc_mgr.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: Клиент создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /clients/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Клиенты]
      summary: Получить клиента
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Клиенты]
      summary: Обновить клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'

    delete:
      tags: [Клиенты]
      summary: Удалить клиента (soft delete)
      description: Только admin.
      responses:
        '204':
          description: Удалён
...
  /equipment:
    get:
      tags: [Оборудование]
      summary: Список оборудования
      parameters:
        - in: query
          name: client_id
          schema:
            type: integer
        - in: query
          name: model_id
          schema:
            type: integer
        - in: query
          name: warranty_status
          schema:
            $ref: '#/components/schemas/WarrantyStatus'
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/EquipmentResponse'

    post:
      tags: [Оборудование]
      summary: Добавить оборудование
      description: admin, svc_mgr.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
      responses:
        '201':
          description: Оборудование добавлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /equipment/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Оборудование]
      summary: Карточка оборудования
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Оборудование]
      summary: Обновить оборудование
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'

    delete:
      tags: [Оборудование]
      summary: Удалить (soft delete)
      description: Только admin.
      responses:
        '204':
          description: Удалено
...
  /equipment/{id}/history:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Оборудование]
      summary: История ремонтов и ТО (UC-1002)
      parameters:
        - in: query
          name: work_type
          schema:
            type: string
            enum: [repair, maintenance]
        - in: query
          name: date_from
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    work_type:
                      type: string
                    work_date:
                      type: string
                      format: date
                    engineer_name:
                      type: string
                    description:
                      type: string
                    ticket_number:
                      type: string
                      nullable: true
...
  /reports/equipment:
    get:
      tags: [Отчёты]
      summary: Отчёт по парку оборудования (UC-1006)
      description: director, svc_mgr, admin.
      parameters:
        - in: query
          name: client_id
          schema:
            type: integer
        - in: query
          name: model_id
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object

  # ═══════════════════════════════════════════════════════════════
  # AUDIT
  # ═══════════════════════════════════════════════════════════════
...